move db creation file to delegated_dbhost
This change will retrieve the file used to initialize the database and then copy it to the `delegated_dbhost` before loading it.  it intentionally does not remove the file afterward, in case something goes wrong and the file needs to be re-imported. it also only executes for the series of conditions around zabbix `>=3.0` and was only tested on Ubuntu.

it resolves #96.